requirements.txt
# --- Bedrock AgentCore & SDK ----------------------------------
bedrock-agentcore==0.1.1
bedrock-agentcore-starter-toolkit==0.1.2
boto3==1.40.0
botocore==1.40.0

# --- Strands --------------------------------------------------
strands-agents[a2a]>=1.4.0
strands-agents-tools[a2a]>=0.2.3
a2a-sdk>=0.2.16

# --- Web サーバ（AgentCore が内部で ASGI 起動） --------------
uvicorn==0.34.2

# --- fastAPI --------------
fastapi==0.116.1
